refactor: AbstractController을 상속하도록 기존 Controller 구조 변경
