+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -5,19 +5,19 @@ import com.techcourse.exception.UserException;
 import com.techcourse.model.User;
 import java.io.IOException;
 import java.net.URISyntaxException;
-import org.apache.coyote.http11.FileReader;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.HttpStatusCode;
 
-public class RegisterController {
+public class RegisterController extends AbstractController {
 
     private static final RegisterController instance = new RegisterController();
 
     private RegisterController() {
     }
 
-    public void register(HttpRequest request, HttpResponse response) throws URISyntaxException, IOException {
+    @Override
+    public void doPost(HttpRequest request, HttpResponse response) throws URISyntaxException, IOException {
         String account = request.getRequestBodyValue("account");
         String email = request.getRequestBodyValue("email");
         String password = request.getRequestBodyValue("password");
@@ -32,6 +32,12 @@ public class RegisterController {
         setResponseContent(request, response);
     }
 
+    @Override
+    public void doGet(HttpRequest request, HttpResponse response) throws URISyntaxException, IOException {
+        response.setHttpStatusCode(HttpStatusCode.OK);
+        setResponseContent(request, response);
+    }
+
     private void checkDuplicatedUser(User user) {
         String userAccount = user.getAccount();
         InMemoryUserRepository.findByAccount(userAccount)
@@ -40,21 +46,9 @@ public class RegisterController {
                 });
     }
 
-    private void redirect(HttpResponse response, String path) {
-        response.setHttpStatusCode(HttpStatusCode.FOUND);
-        response.setHttpResponseHeader("Location", path);
-    }
-
     private void setFailResponse(HttpRequest request, HttpResponse response) {
         request.setHttpRequestPath("/register.html");
         response.setHttpStatusCode(HttpStatusCode.BAD_REQUEST);
-    }
-
-    private void setResponseContent(HttpRequest request, HttpResponse response) throws URISyntaxException, IOException {
-        FileReader fileReader = FileReader.getInstance();
-        response.setHttpResponseBody(fileReader.readFile(request.getHttpRequestPath()));
-        response.setHttpResponseHeader("Content-Type", request.getContentType() + ";charset=utf-8");
-        response.setHttpResponseHeader("Content-Length", String.valueOf(response.getHttpResponseBody().body().getBytes().length));
     }
 
     public static RegisterController getInstance() {
