+++ tomcat/src/main/java/com/techcourse/controller/StaticPageController.java
@@ -2,45 +2,21 @@ package com.techcourse.controller;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
-import org.apache.coyote.http11.FileReader;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.session.SessionManager;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.HttpStatusCode;
 
-public class StaticPageController {
+public class StaticPageController extends AbstractController {
 
     private static final StaticPageController instance = new StaticPageController();
 
     private StaticPageController() {
     }
 
-    public void getStaticPage(HttpRequest request, HttpResponse response) throws URISyntaxException, IOException {
+    @Override
+    public void doGet(HttpRequest request, HttpResponse response) throws URISyntaxException, IOException {
         response.setHttpStatusCode(HttpStatusCode.OK);
-        if (request.getHttpRequestPath().equals("/login") && checkLogin(request)) {
-            redirect(response, "/index.html");
-        }
         setResponseContent(request, response);
-    }
-
-    private boolean checkLogin(HttpRequest request) {
-        String jsessionid = request.getJSESSIONID();
-        if (jsessionid.isEmpty()) {
-            return false;
-        }
-        return SessionManager.containsSession(jsessionid);
-    }
-
-    private void redirect(HttpResponse response, String path) {
-        response.setHttpStatusCode(HttpStatusCode.FOUND);
-        response.setHttpResponseHeader("Location", path);
-    }
-
-    private void setResponseContent(HttpRequest request, HttpResponse response) throws URISyntaxException, IOException {
-        FileReader fileReader = FileReader.getInstance();
-        response.setHttpResponseBody(fileReader.readFile(request.getHttpRequestPath()));
-        response.setHttpResponseHeader("Content-Type", request.getContentType() + ";charset=utf-8");
-        response.setHttpResponseHeader("Content-Length", String.valueOf(response.getHttpResponseBody().body().getBytes().length));
     }
 
     public static StaticPageController getInstance() {
