+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,12 +1,10 @@
 package org.apache.coyote.http11;
 
-import com.techcourse.exception.UncheckedServletException;
+import com.techcourse.controller.Controller;
 import java.io.BufferedReader;
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.net.Socket;
-import java.net.URISyntaxException;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.request.HttpRequestParser;
@@ -38,12 +36,12 @@ public class Http11Processor implements Runnable, Processor {
             String response = getResponse(inputStream);
             outputStream.write(response.getBytes());
             outputStream.flush();
-        } catch (IOException | UncheckedServletException | URISyntaxException | IllegalArgumentException e) {
+        } catch (Exception e) {
             log.error(e.getMessage(), e);
         }
     }
 
-    private static String getResponse(InputStream inputStream) throws IOException, URISyntaxException {
+    private static String getResponse(InputStream inputStream) throws Exception {
         RequestMapper requestMapper = RequestMapper.getInstance();
         HttpRequestParser httpRequestParser = HttpRequestParser.getInstance();
         HttpResponseParser httpResponseParser = HttpResponseParser.getInstance();
@@ -51,7 +49,8 @@ public class Http11Processor implements Runnable, Processor {
 
         HttpRequest httpRequest = httpRequestParser.parseRequest(bufferedReader);
         HttpResponse httpResponse = new HttpResponse();
-        requestMapper.mapRequest(httpRequest, httpResponse);
+        Controller controller = requestMapper.mapRequest(httpRequest, httpResponse);
+        controller.service(httpRequest, httpResponse);
         return httpResponseParser.parseResponse(httpResponse);
     }
 }
